key assets metadata by file name rather than fetch URL location (#99)
* abstract asset directory file ops / magic values

* key assets metadata by file name rather than fetch URL location

Assets locations change per environment and we want to re-fetch any
assets which differ in case of an environment change. Keying the
asset metadata by the output file name ensures a correct etag
comparison (comparing etag of the local file vs the remote origin)
in case of an environment change.