Load assets once when running N tests in parallel, rather than N times (#172)
We're intermittently getting failures like:

```
[run-all-tests] Starting simulation...
[run-all-tests] 
[run-all-tests] ERROR: TypeError: replayUserInteractions is not a function
[run-all-tests] ERROR:     at replayEvents (/home/runner/work/meticulous/meticulous/node_modules/@alwaysmeticulous/replayer/dist/replayer.js:107:11)
[run-all-tests] ERROR:     at async replayCommandHandler (/home/runner/work/meticulous/meticulous/node_modules/@alwaysmeticulous/cli/dist/commands/replay/replay.command.js:117:5)
[run-all-tests] ERROR:     at async handleReplay (/home/runner/work/meticulous/meticulous/node_modules/@alwaysmeticulous/cli/dist/deflake-tests/deflake-tests.handler.js:14:20)
[run-all-tests] ERROR:     at async deflakeReplayCommandHandler (/home/runner/work/meticulous/meticulous/node_modules/@alwaysmeticulous/cli/dist/deflake-tests/deflake-tests.handler.js:35:25)
[run-all-tests] ERROR:     at async main (/home/runner/work/meticulous/meticulous/node_modules/@alwaysmeticulous/cli/dist/parallel-tests/task.handler.js:46:20)
[run-all-tests] ERROR: 
[run-all-tests] Simulation done!
[run-all-tests] 
```

And:

```
[run-all-tests] ERROR: TypeError: setupReplayNetworkStubbing is not a function
```

(more details in SIM-308)

I think it's likely that this is because of a race condition in loading
the assets. If you start 10 tests at once in 10 threads, then they'll
all load assets at the same time (etag in metadata only updated after
asset finishes loading), and try writing to the same files. This PR
pulls the asset loading up, so it only happens once before running any
of the tests, instead of one per test. If this doesn't solve the issue
(we still see such failures) then at least it will have made the code
more performant.

Testing performed:
 - run-all-tests still works
 - replay still works