Add baseCommitSha argument (#174)
This should enable a workflow where the baseReplayIds are not checked
into your meticulous.json file. So if the diffs are intended you can
just merge the PR, without having to update your meticulous.json file.
The intent is that this will be used as the recommended method for both
active and passive tests. The --baseCommitSha will be automatically
passed in by a github action that wraps the Meticulous CLI.