Retry on failure (#186)
Adds the new argument:

   **name**: maxRetiresOnFailure
**description**: If set to a value greater than 0 then will re-run any
replays that give a screenshot diff and mark them as a flake if the
screenshot generated on one of the retryed replays differs from that in
the first replay.

And stores such flakes in the backend as:

```ts
export interface ScreenshotDiffResultFlake {
  outcome: "flake";

  /**
   * The original diff. Can be any outcome but for 'no-diff'.
   */
  diffToBaseScreenshot: Omit<ScreenshotDiffResult, "identifier">;

  /**
   * The diffs created by retrying taking the head screenshot and comparing
   * it to the original head screenshot. At least one of these will have an
   * outcome other than no-diff, hence why this failure is marked as a flake.
   *
   * Note that in the context of these diffs base means the original head screenshot taken,
   * and head means the new head screenshot taken.
   */
  diffsToHeadScreenshotOnRetries: Array<
    Omit<
      ScreenshotDiffResult | { outcome: "missing-base-and-head" },
      "identifier"
    >
  >;
}
```

Currently the UI completely ignores flakes ([example test
run](https://app.meticulous.ai/projects/meticulous/meticulous_internal_tests/test-runs/ttDhDDFCqHdm8GftMkgDpqDfth)),
though after this PR I'll update it to provide a clearer message.
Something like:

> **All 1 screens tested had zero reproducible differences**
> Meticulous replayed 1 user sessions before and after your PR and
compared the 1 screenshots taken in each run. Differences were found for
1 of the screenshots, but when we retried taking those screenshots we
got different results - so we're assuming the original difference was
just a flake.

Or in the case there are some identical screenshots too, then:

> **All 1 screens tested had zero reproducible differences**
> Meticulous replayed 1 user sessions before and after your PR and
compared the 2 screenshots taken in each run. Differences were found for
1 of the screenshots, but when we retried taking those screenshots we
got different results - so we're assuming the original difference was
just a flake.