Better cli error handling (#190)
Currently when the code throws an Error the error is never logged out
(the code in index.ts didn't seem to be working), so the command failed
with no feedback as to why. This updates it to log out the error. I also
updated to stop it from posting the big block of help text which can
make it hard to find and read the error. Note that if you pass a command
or argument that doesn't exist, miss a required arg, pass the wrong type
to an arg etc. then it'll still show the help text. But not if the
command handler itself throws an arbitrary error.

Side question:
- Do we want to use `throw new Error(...)` or `process.exit` as our
standard pattern? Or a mix? Note that `throw new Error(...)` means (1)
it's logged to Sentry (may or may not be desired for different error
types), (2) it's will execute any finally calls or catch handling, not
just in the direct function the error is thrown, but in any parent
function that calls that function. Normally this is desired. Maybe we
lean towards `throw new Error(...)` and if we don't want all Errors to
be logged to Sentry, we introduce some special types of errors like
`InvalidCliArgumentsError` that doesn't get Sentry logged (or logged
differently) - i.e. seperate out errors likely due to user error vs due
to bug in our code?