Do not delete .lock-target directories (#197)
Keeping the directories around was causing errors, since a second
process would delete the lock-target in-between when a first process
created it and started trying to acquire the lock:

```
[Error: ENOENT: no such file or directory, lstat '/github/home/.meticulous/sessions/<redacted>.json.lock-target'] {
  errno: -2,
  code: 'ENOENT',
  syscall: 'lstat',
  path: '/github/home/.meticulous/sessions/<redacted>.json.lock-target'
}
```

This should fix these errors. If we find the .lock-target files hanging
around distracting then we could later switch to a different technique.