Add essential features only flag (#206)
When debugging it's often using to disable things like rrweb recording,
since it adds a lot of noise to the logs (adding MutationObservers,
scheduling timeouts etc.). Filtering it out enables you to focus on just
the execution of the core application, and debug that. This PR adds a
flag to disable an extraneous features, kind of like 'Safe Mode' for
Windows 98.