Default --skipPauses to true (#229)
This is now the best tested and most reliable way to run the tests (e.g.
notice that `--skipPauses false` tests flake a lot on Meticulous CI, but
`--skipPauses true` tests do not), so we should default it to true so
users get the best experience by default. I'll leave in `--skipPauses
false` as a debugging option.