Per test-case appUrl support (#302)
### Changes

This PR fixes the CLI support for specifying `appUrl` per test case via
the `replay option`. Type-wise it was already supported though we didn't
consume the `appUrl` correctly, this fixes it and also updates the
replay target generation to take into account the replay option `appUrl`
if set.

Have made `--appUrl` and the per-test-case `appUrl` mutually-exclusive
as it can lead to priority confusion but do shout it you think there are
cases where we should allow doing this!