Expand list of test cases to run with those from the base test run (#340)
This PR updates the `runAllTests` logic to expand the list of test cases
to include those from the base test run. This is necessary to prevent a
drop in coverage when the project config test cases are updated. Such
updates have the potential to entirely replace the set of sessions to
run and will occur more frequently in the future.