Auto-generate logs.concise.txt (#419)
When debugging replays logs.json can be pretty verbose, so we often use
jq to create a more readable version. This is time consuming so this PR
updates the CLI to generate it automatically. We could generate it at
replay time, but that'd increase the size of the replay zips. So better
to generate at consumption time.

Potentially we may want Meticulous BE to generate this instead, and
display it in the web-app. However it's not obvious how often the
internal logs will be useful for users. I expect we'll want to do this
at some point though with the raw console logs. If when we do that we
may potentially be able to remove this code here, and add it to the
Meticulous FE/BE instead.