Simplify recorder script injection (#552)
Simplifies the injection of recorder scripts by injecting
'v1/meticulous-manual-init.js' and only initializing in intended frames,
instead of installing the early network recorder, and then later only
injecting 'v1/meticulous.js' in intended frames.

Note that `.on('frame_navigated', frame => frame.evaluate(...))` does
not guarantee that the injected code is executed before any other code
on the page (and in practice often evaluates it after), which is why the
early network recorder used to be required. In contrast
evaluateOnNewDocument does provide this guarantee.

This simplification also works around a potential bug in Chrome for
Testing, where the combination of the 'v1/meticulous.js' AND the early
network recorder javascript AND Chrome for Testing (instead of the older
versions of Chrome used by older Puppeteer) lead to an 'Aw Snap!' error
at the point in time when the early network recorder javascript is
injected into some sub-iframes (for example when navigating from
google.com to google images). This bug does not happen when one of these
three elements is not present, and has not been fully root caused. Full
details in Linear issue ENG-803.