Add early network recorder loader method (#554)
See jsdoc comment for when this is useful. I don't know how common this
will be / whether worth supporting, but feels like it will continue to
come up?

It will also allow us to have a more performant production recording
setup if we can get the early network recorder snippet to be smaller
(currently it's 0.5mb, while full snippet is 1.2mb).