Fix record login flow  (#556)
Previously would always give a 'Invalid arguments: should be exactly one
string' on Meticulous initialization. I think you get this if you call a
Puppeteer exploreFunction while it's still initialising, and the
evaluateOnNewDocument must have been happening while
`__meticulous_onBeginRecording` was being attached or something. This
shifts the ordering to ensure the methods are guaranteed to be bound
before they are called.