Create tunnel client (#590)
### Changes

This adds a client for our secure tunnels service.

This client is based on https://github.com/localtunnel/localtunnel with
some modifications to make it secure and tailor it for our use case.

This package expose a command for interacting with the client but after
the initial testing most of the access will be programatically via the
API, e.g. within GitHub Actions or as part of other command handlers.
The API will most likely change in following versions as things are
still in developmens.

### Code review

Review by commit might be easier, first PR is quite large though but
it's just forking the upstream master.