Don't download source maps for non-JS files (#640)
In an upcoming PR, the snippet server will also start serving a WASM
file for parsing sourcemaps. This won't come with a sourcemap so the
download for a `.map` file below will fail. To avoid a bunch of retries
and 1 minute wait, check the extension is `.js` before we try to
download a sourcemap.