Connect to specific tunnel server when establishing tunnel connections
Previously we used `tunnels.metiuclous.ai` to establish tunnel connections
rather than `tunnel-X.tunnels.metiuclous.ai`. This is okay at the moment
as both point to the same host but won't be the case in the future.

This change is temporary and assumes that the tunnel URL is a subdomain
of the correct host to connect to (e.g `tunnel-1.meticulous.ai` for a
tunnel with URL `slug.tunnel-1.meticulous.ai`) which is the case currently. The
host and scheme will become BE-driven in future versions