Add support for non-connection pooling multiplexing secure tunnels
This creates a new 'multiplexing cluster' for handling tunnels
in 'non-pooling' connection mode (the 'cluster' naming here is a bit
confusing as this no longer deals with a connection pool but can
be renamed in the future).

In this new non-pooling mode new (logical) connections are opened from
the secure tunnel server on each incoming request.