Improved docs regarding deduped events (#671)
The debugger UI deals with event indexes after deduplication (i.e. gets
given an event list after deduplication, and sends commands to advance
to indexes in the deduplicated list), but we display event numbers to
the user as pre-deduplication event numbers, so that they correspond to
what they see in the session data.