Drop Axios for API requests (#913)
As requested by our customer who needs to run behind a proxy, because we
know `node-fetch` plays nicer with proxies. I have not yet ripped Axios
out completely as it's still in `download-file.ts`. I believe this is
fine because none of the flows this customer needs involve files being
downloaded in their CI runners. However, just in case I've structured
the code so it should be easier to rewrite that part too (in particular,
I've exported the methods in `client.ts` that do stuff like wrapping a
request with retries, which will be needed there).