Better handling of symlinks in asset uploads (#946)
The original implementation had some issues with how it handled loops
due to inconsistencies between relative and absolute paths which could
lead to infinite recursion. I've rewritten this to not be recursive,
handle loops correctly, and be more explicit throughout about what is a
relative path and what is an absolute one.