Add timeout for zip extraction (#965)
Retrying is tricky because we can't stop the original promise which
might still be trying to output to the file. I think it'd be a bigger
rewrite to handle that properly. Instead, let's just fail hard and let
the caller handle the retrying (in the cause of cloud replay, this will
bubble all the way up to Argo which will restart the chunk).