Add PostProcessing status to test run (#976)
The intent is to use this so that session selection can start
downloading the coverage of replays without having to wait for the
post-processing step to run. For a large customer the post-processing
step can take a couple of minutes (we should try to speed this up), and
since we do 3 runs in session selection that can add up to 5-10 minutes
over the whole run.