[Feature Proposal] Add cookie redaction middleware (#991)
We need to redact specific cookie values for our implementation, and
figured it would be better to upstream this.

Notes:
* The intention is a calling pattern of `middleware: [
redactCookies(['sessionId', 'foo']) ]`
* The middleware looks at `cookie` and `set-cookie` headers
* It treats everything as case-insensitive but tries to preserve the
existing case in the output (this seems most spec-compliant, although as
with many things like this the Wild West is wild)
* The redaction asterisks length intentionally does NOT match the
original length, to avoid leaking data (extremely unlikely obviously,
but this seemed cleaner to me)

I let Cursor do some of this work and write the tests, although I
reviewed everything fairly carefully. But if there's anything bizarre in
there let me know.

---------

Co-authored-by: Giles Copp <gilesc@dropbox.com>